requirements.txt
# SportsHub backend – minimal for deployment
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
redis>=4.0
gunicorn>=21.0
dj-database-url>=2.0
python-dotenv>=1.0
psycopg2-binary>=2.9
boto3>=1.28
django-storages>=1.14
